+++ db/migrate/20161227121934_add_index_to_projects_project_types.rb
@@ -17,13 +17,13 @@
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 
 class AddIndexToProjectsProjectTypes < ActiveRecord::Migration
-  def self.up    
-    add_index :projects_project_types, :project_id, unique: true
-    add_index :projects_project_types, :project_type_id
+  def self.up   
+    add_index :projects_project_types, :project_id, unique: true unless index_exists?(:projects_project_types, :project_id)
+    add_index :projects_project_types, :project_type_id unless index_exists?(:projects_project_types, :project_type_id)
   end
   
   def self.down
-    remove_index :projects_project_types, :project_id
-    remove_index :projects_project_types, :project_type_id
+    remove_index :projects_project_types, :project_id if index_exists?(:projects_project_types, :project_id)
+    remove_index :projects_project_types, :project_type_id if index_exists?(:projects_project_types, :project_type_id)
   end
 end
